+++ preliz/ppls/__init__.py
@@ -1,0 +1,4 @@
+from .agnostic import posterior_to_prior
+
+
+__all__ = ["posterior_to_prior"]
